requirements.txt
numpy==1.26.4
pandas==2.2.1
seaborn==0.13.2
scikit-learn==1.4.1post1
torch==2.2.1
matplotlib==3.8.3
pyyaml==6.0.1